seperate app from server.ts
