+++ src/__tests__/profile.controller.test.ts
@@ -1,0 +1,2 @@
+import request from 'supertest'
+import app from '../server'
