+++ src/server.ts
@@ -2,11 +2,9 @@ import dotenv from 'dotenv'
 import express, { Request, Response } from 'express'
 import cors from 'cors'
 import helmet from 'helmet'
-import logger from './lib/logger'
 import mongo from './config/mongo.config'
 import morgan from './config/morgan'
 import compression from 'compression'
-import chalk from 'chalk'
 import ProfileRouter from './routes/profile.route'
 import UserRouter from './routes/user.route'
 import LogsRouter from './routes/logs.route'
@@ -31,14 +29,7 @@ app.use(morgan)
 
 app.use('/api', ProfileRouter, UserRouter, LogsRouter, HealthCheckRouter)
 app.get('/', (req: Request, res: Response) => {
-  res.send('digimon api')
+  res.send('digimon api service')
 })
 
-const PORT: number = parseInt(process.env.PORT as string, 10)
-
-app.listen(PORT, () => {
-  logger.info(
-    chalk.bgCyan(`[デジタルモンスター]`),
-    chalk.bold.blue(`: running on http://localhost:${PORT}`)
-  )
-})
+export default app
